pyproject.toml
[project]
name = "agentic-patterns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph>=0.0.10",
    "langchain-openai>=0.1.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "grandalf>=0.8",
]
